Fixed transient disconnect issue
So... we define being 'disconnected' as
having not connected to gmail after n secs,
but if when it finally does reconnect, gmail
hasn't updated its 'modified' field, then
onUpdate will never be called. Now we track
that we've called onDisconnect and call
onUpdate the first time a succesful connection
is made, regardless of whether there's actually
any new information to report.

Not sure this is the best idiom for gConn as
an object, but it is convenient for the application,
and will be used for now.